requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
av==16.1.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
cryptography==46.0.5
ctranslate2==4.7.1
distro==1.9.0
dotenv==0.9.9
exceptiongroup==1.3.1
faster-whisper==1.2.1
filelock==3.24.3
flask==3.1.3
flatbuffers==25.12.19
fsspec==2026.2.0
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
googleapis-common-protos==1.72.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface-hub==1.5.0
humanfriendly==10.0
idna==3.11
itsdangerous==2.2.0
jinja2==3.1.6
jiter==0.13.0
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
numpy==2.4.2
oauthlib==3.3.1
onnxruntime==1.24.2
openai==2.24.0
opencv-python==4.13.0.92
packaging==26.0
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1-modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pygments==2.19.2
pyparsing==3.3.2
pyreadline3==3.5.4
python-dotenv==1.2.2
pyyaml==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.3
rsa==4.9.1
setuptools==82.0.0
shellingham==1.5.4
sniffio==1.3.1
sympy==1.14.0
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.1
typing-extensions==4.15.0
typing-inspection==0.4.2
uritemplate==4.2.0
urllib3==2.6.3
werkzeug==3.1.6
yt-dlp==2026.2.21
